requirements.txt
discord.py>=2.3.2
PyNaCl>=1.5.0
yt-dlp>=2024.03.10
spotipy>=2.23.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
